+++ src/test/java/com/fgiannesini/neuralnetwork/activationfunctions/IdentityFunctionApplierTest.java
@@ -1,6 +1,6 @@
 package com.fgiannesini.neuralnetwork.activationfunctions;
 
-import org.jblas.FloatMatrix;
+import org.jblas.DoubleMatrix;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -18,28 +18,28 @@ class IdentityFunctionApplierTest {
     @Test
     void apply() {
         Assertions.assertAll(
-                checkTanhApply(new float[]{0}, new float[]{0}),
-                checkTanhApply(new float[]{1f}, new float[]{1}),
-                checkTanhApply(new float[]{-1f}, new float[]{-1}),
-                checkTanhApply(new float[]{1f, 2f, 3f, 4f, 5f}, new float[]{1, 2, 3, 4, 5})
+                checkTanhApply(new double[]{0}, new double[]{0}),
+                checkTanhApply(new double[]{1f}, new double[]{1}),
+                checkTanhApply(new double[]{-1f}, new double[]{-1}),
+                checkTanhApply(new double[]{1f, 2f, 3f, 4f, 5f}, new double[]{1, 2, 3, 4, 5})
         );
     }
 
-    private Executable checkTanhApply(float[] expected, float[] input) {
-        return () -> Assertions.assertArrayEquals(expected, identityFunctionApplier.apply(new FloatMatrix(input)).data, 0.0001f);
+    private Executable checkTanhApply(double[] expected, double[] input) {
+        return () -> Assertions.assertArrayEquals(expected, identityFunctionApplier.apply(new DoubleMatrix(input)).data, 0.0001f);
     }
 
     @Test
     void derivate() {
         Assertions.assertAll(
-                checkTanhDerivate(new float[]{0f}, new float[]{0}),
-                checkTanhDerivate(new float[]{2f}, new float[]{2}),
-                checkTanhDerivate(new float[]{-2f}, new float[]{-2}),
-                checkTanhDerivate(new float[]{1f, 2f, 3f, 4f, 5f}, new float[]{1, 2, 3, 4, 5})
+                checkTanhDerivate(new double[]{0f}, new double[]{0}),
+                checkTanhDerivate(new double[]{2f}, new double[]{2}),
+                checkTanhDerivate(new double[]{-2f}, new double[]{-2}),
+                checkTanhDerivate(new double[]{1f, 2f, 3f, 4f, 5f}, new double[]{1, 2, 3, 4, 5})
         );
     }
 
-    private Executable checkTanhDerivate(float[] expected, float[] input) {
-        return () -> Assertions.assertArrayEquals(expected, identityFunctionApplier.derivate(new FloatMatrix(input)).data, 0.0001f);
+    private Executable checkTanhDerivate(double[] expected, double[] input) {
+        return () -> Assertions.assertArrayEquals(expected, identityFunctionApplier.derivate(new DoubleMatrix(input)).data, 0.0001f);
     }
 }
